Able to reorder points
The classes boundary, path, node, and box now have methods that return
the number of points based on half the number of coordinates stored. The
classes boundary, node, and box have a way to reorder the saved points
going counterclockwise, starting in the lower-right corner, and
repeating the initial point. All pointers for variable-length
coordinates in a class and the class properties are removed since they
are no longer needed when using vectors of GDSII elements in a single
cell. A more terse printAlt() function was added based off what Li
requested.