SPEF file writes extracted design parasitics
The simulation input file was rewritten to conform to the length that Li
needs and contain the fields that Michael needs. The simulation input
file may use either "numPorts" or "numPort". The main function puts
the fdtdMesh variable field containing Y parameters into matrices needed
to write the SPEF file.