SolverDataBase method initializes fdtdMesh
A function SolverDataBase::convertToFDTDMesh() was written that takes
existing fields in the SolverDataBase to start a fdtdMesh variable
without performing any of the meshing contained in readInput() found in
mesh.cpp. The function will eliminate a file read, but it is not used
anywhere yet. A function will be needed for the meshing part later.