Incomplete use of convertToFDTDMesh
Modified fdtd.h and mesh.cpp to have readInput() become meshAndMark() so
that the stack file is no longer needed. Unnecessary variables in the
function were deleted. Now TestMain.cpp uses the statuses returned by
the readers to change the error message, and the flow of the whole
simulation was adjusted. Modified solnoutclass.h to fix an error where
the lengthUnit was not being used for layers or ports. *Warning*: the
code produces segmentation faults for unknown reasons following the code
consolidation. SPEF files cannot be written from IC designs.