Facebook Wedge 100 and Corrections
A GDSII file and simulation input file were added based on the Facebook
Wedge 100 in the Open Compute Project. The files added were produced
manually after edits to the Gerber files and transformations in Klayout.
An index decrementing error was corrected in the reorder method of the boundary, box, and node classes. The SolverDataBase readSimInput() method was improved by allowing comments to be used with ports when given layer numbers.