Changed Facebook stack file
The stack file was modified so that no comments appear. The file
mesh.cpp had printing added to locate a segmentation fault. Part of the
error is that the PCB has a stack file that starts with the top layer
first instead of the bottommost layer in the stackup. The solution is to
have Li add code that takes the z-coordinate of the bottom of each layer
to have design-file paradigm independence.