Type-related compilation errors fixed
All compiler errors related to class definitions and types were fixed.
TestMain.cpp had the boolean variables for --simulate mode initialized
at the top, and markAndMesh() was fully embraced with clock object t2
monitoring the time. There were multiple declarations of the same
objects in fdtdMesh class in fdtd.h that were reduced to a single
declaration. The print() function for fdtdMesh class implemented in
mesh.cpp was modified for the present objects of the class.