Removed Stack Files and Fixed adb.print()
Brought simulation input files in line with what Li was using in the stack files for the examples. An error was fixed in limboint.h which required at least one cell to be printed in order for conductor information to be saved to the mesh (now handled in TestMain.cpp). Removed many print lines from the code. In all places, the custom "doublecomplex" was replaced with C++ "complex<double>". Currently, Paradiso for solving the reference matrix is commented off to be faster.