README.md Adds Simulation Input Syntax
The flowchart image in README.md was updated to show the Xyce file
output. README.md was updated with a new section giving the simulation
input file syntax, comment rules, and acceptible values for
placeholders. Updates were made to "solnoutclass.h" so that
SolverDataBase::readSimInput() could actually handle block-interrupting
comments, in-line comments on all important lines, and linear frequency
interpolation of the sweep.