Gerber Outline File Reading in Place
A new class for the photoplotter apertures which Gerber files can
specify was added. A crude but exacting method was added to
SolverDataBase that parses the needed Gerber file elements to draw the
outline of a PCB was added. The fidelity has not been verified. A convex
hull algorithm might be the easiest way to go from graphical elements of
an outline to the coordinates of the path of the design limits.