Convex Hull Method Added
The back drill Gerber file serving as the board outline for the Facebook
Wedge 100 was added. The header limboint.h had a structure added to
perform the comparison needed to sort points for the Graham scan convex
hull agorithm in addition to AsciiDataBase::convexHull() to implement
the method. The header solnoutclass.h had the
SolverDataBase::readGerberOutline() method switched to output
vector<double> and a more careful name of the output GDSII file for the
board outline. The file-scope variables of that method and
readIMPwriteGDSII() were made effective by placing them outside the
scope of line-by-line reading. Gerber file reading of CW or CCW circular
interpolation was corrected by making sure points were placed along a
24-gon, not nArcPt-gon. The method readSimInput() was updated to allow
a Gerber file name and the z-direction limits instead of all six limits
on the IC design size line. Note that a method of passing the boundary
to fdtdMesh must be agreed upon because right now NaN is used for
x-direction and y-direction limits, preventing a PCB run.