More Detailed singleStrip Run
Additional information from the IMP file was added to
singleStrip.sim_input. Errors are certain to occur if the values of zmin
and zmax do not correspond exactly to the bottom layer and top layer in
the stack-up. HYPRE will hang if the discretization is too fine. The
lines in hypreSolve.cpp were experimented on to discover what can cause
HYPRE to exit with false success. Printing to terminal was updated to
C++ streams with timing reported with the seconds unit.

Previous commit updated .gitignore for convenience. The file
TestMain.cpp was commented for parrot control mode to deactivate the ad
hoc production of PSLG files based on specific examples. Simulation
input files were updated to give a dielectric margin on the edges of
each layer. New preprocessor macros in fdtd.h control the HYPRE solver
settings. A function prototype was moved to hypreSolver.h. The file
hypreSolve.cpp was updated for appearance when diagnosing the pernicious
false success returned issue. There was bad logic when reordering
boundary, node, or box outline points in limboint.h to CCW that was
corrected along with GeoCell::printAlt(). Comment checking in
SolverDataBase::readSimInput() was altered to handle EOF.