Interfaces GDSII Files with AREF and STRANS
The header limboint.h had a helper class "strans" added to handle
x-direction mirroring, magnification, and rotation. For these specific
operations, logic was added to determine how one transformation composes
with another. The class also can apply the transformation to a single
point. The GDSII storage classes "textbox", "sref", and "aref" have an
object of type "strans". The entire "aref" class was added to handle the
seventh and final GDSII element that remained. It remains to be seen how
well "aref" handles non-Manhattan arrays through findColSep() aand
findRowSep() methods in the useful findInstances(). A new object for
"aref" was added to "GeoCell", and print() and printAlt() were updated.
All recursive methods for cell references were updated to use
size-checked, length-2 vectors for the translations and the new "strans"
class for linear transformations. Extensive changes to saveToMesh() have
updated the comments, changed the iteration logic to use fewer lines of
code, and handle path termination types. The necessary changes to the
general callback function for the reader and dump() were made. The
header solnoutclass.h had changes made to setPortsGDSII() corresponding
to how recursion is handled.