Update the algorithm of finding the nodes inside polygons.
(1) Update the algorithm of fast finding the nodes inside polygons. Now it can deal with arbitrary shapes instead of with boundaries only along x or y axis. (2) Update the V0d2 generation. Now it doesn't use unordered_map, which is slow when the size becomes large. The generation of V0d2 is based on the analytical expression of V0d and V0da. (3) Update the reading algorithm to be applicable to the z coordinates not started with 0.