All Ports Located for SDFFRS_X2 and 4004
More comments were updated to report time in seconds. "TestMain.cpp" was
updated so that CITI files are always dumped with S-parameters. The
simulation input files recognize 10 ports for "SDFFRS_X2.gds" and 16
ports for "4004.gds". A SPICE transient analysis file was added ad hoc
for Xyce use with "SDFFRS_X2.cir". Newlines were reduced in the terminal
output, and the printing in "hypreSolve.cpp" and "matrixCon.cpp" was
compacted. New methods were added to Parasitics in "solnoutclass.h" to
convert between any two of scattering, admittance, or impedance
parameters in addition to terminating with a known ZLoad or YLoad.

Always remember to check how disMin is calculated when the most recent
commit switches between Li and Michael for "mesh.cpp".