Small Changes for GDSII Packages
The preprocessor directive for MINDISFRACXY was made a less aggressive 1e-4
instead of 1e-6, which should still capture down to 2 mils on the Facebook
Wedge 100 based on its dimensions. This prevents the server from running out
of memory immediately when running the IBM whole package edge calculations.
The Cadence BRD-to-Stream converter used the FORMAT element in the GDSII
file, so an inactive statement was added to handle that.