Added DataType to Boundaries and Paths
The classes "boundary" and "path" in limboint.hpp were updated to have a
datatype object and the associated constructor, getter, and setters functions.
The long print function for the "GeoCell" class was updated accordingly. The
method AsciiDataBase::saveToMesh() was updated to skip the identification of
boundaries, paths, and boxes as conductors if the datatype for the former two
or the box type for the last one is not zero.