Method to Read Excellon NC Drill File
A new method SolverDataBase::readHoles() was written based off the previous
method to read the outline Gerber file that returns an AsciiDataBase with
drill hole information matching the IPC-NC-349 or XNC or Excellon format.
Note that no control mode of the executable uses this untested code at this
time and that changes to solver conductor identification have yet to be made.