Only Pass Valid Layers to Solver
The solver class fdtdMesh was updated to have different conductivities for
the conductor materials within the layer rather than copper only. The new
conductivity information is read in but not used anywhere.

Some reordering in TestMain.cpp was needed to exchange information about
layers the solver should ignore in AsciiDataBase::saveToMesh() and
SolverDataBase::convertToFDTDMesh(). The Layer class has new method
Layer::isValid() to check the height, epsilon_r, and sigma for ability to
manufacture in physical reality. The SolverDataBase class has methods to
return the GDSII layer number of ignored layers or a vector of only the
layers the solver needs to handle.

A small change to AsciiDataBase::findPoint() ignores the boundaries, paths,
and box outlines with nonzero datatype or nonzero box type.