Cleaned Up HYPRE Calls
Important macros for HYPRE moved from "fdtd.hpp" to "hyperSolver.h".
Comments were added to "hyperSolve.cpp" while reordering the function
calls and removing unused variables. A correction was made to
paraGenerator() to calculate the line integral only over a single port
side. The dead function setHypreMatrix() was removed from
"matrixCon.cpp". An attempt to parallelize with OpenMPI was aborted
because of its poor fit for this code.