Preprocessor for GDSII Datatype and 64-bit Int
Small corrections were made to the installation instructions. The
makefile went back to sequential Intel MKL linker flags. A preprocessor
directive in "limboint.hpp" controls whether or not the datatype != 0
GDSII file elements are treated as physical or not because the Cadence
conversion might not even use datatype == 0. A few variables in
"mesh.cpp" were instantiated as long long int instead of int because it
might be needed. In the same file, an error message was used if the port
node indices could not be found while the brace formatting was
standardized throughout the file.