Update the function of the boundary definition.
Now the port in the .sim_input are formatted as "Excited port coordinates x, y, z. Reference port coordinates x, y, z." Treat the reference conductor's upper and lower plane nodes as in the PEC. Do in this way because sometimes we cannot treat the whole plane as PEC but just the nodes inside conductors as PEC on the upper and lower planes (like plasma_windowed2.gds).

We can also define "UPPER_BOUNDARY_PEC" and "LOWER_BOUNDARY_PEC" to set the whole upper and lower boundaries as PEC.