Accuracy Restored by Fixed Current Setting
Comments explain each matrix-vector multiplication to improve
correspondence with paper detailing algorithm. A dead task in the
paraGenerator() subflow was removed. The accuracy of the network
parameters and consistency across executions was achieved because of the
modification of the current density setting to use pointer "Jport"
inside the scope of each port solve rather than have ports modify
"sys->J" concurrently. Removing this thread safety violation fixes an
issue dogging the developer for a couple months.