Fix an I/O bug from linux to windows
The bug is from different sizes of 1D vectors inside the 2D vector.

After fixing this bug, the Linux and Windows generate exactly the same Z-parameters for singleStrip_100 example, whose sim_input file is also included in this commit.

Another thing is that win32 and win64 give exactly the same results, too. So, now the layeredFD and reference match among linux, win32, and win64.