Support for Custom Path Extensions
Two new objects (i.e., bexten and eexten) to the class Path to support
pathtype = 4 paths with custom end caps. The constructor, getters, setters, and comments were updated accordingly. A new method for Path was added as a parametrized constructor for pathtype = 4 paths with explicity beginning extension and end extension. The PSLG and mesh drawing codes were modified to identify pathtype = 4 paths and interpret overshoot1 and overshoot2 as the specified values in units of half-widths. The Limbo interface reader will identify records with BGNEXTN and ENDEXTN appropriately. Note that Limbo itself does not support writing these two records at the time of this commit.